+++ Examples/FailoverTest.cc
@@ -20,6 +20,7 @@
 #include <sstream>
 
 #include <LogCabin/Client.h>
+#include <LogCabin/Debug.h>
 #include <LogCabin/Util.h>
 
 namespace {
@@ -36,16 +37,19 @@ class OptionParser {
         : argc(argc)
         , argv(argv)
         , cluster("logcabin:5254")
+        , logPolicy("")
         , timeout(LogCabin::Client::Util::parseDuration("10s"))
     {
         while (true) {
             static struct option longOptions[] = {
                {"cluster",  required_argument, NULL, 'c'},
-               {"timeout",  required_argument, NULL, 't'},
                {"help",  no_argument, NULL, 'h'},
+               {"timeout",  required_argument, NULL, 't'},
+               {"verbose",  no_argument, NULL, 'v'},
+               {"verbosity",  required_argument, NULL, 256},
                {0, 0, 0, 0}
             };
-            int c = getopt_long(argc, argv, "c:t:h", longOptions, NULL);
+            int c = getopt_long(argc, argv, "c:t:hv", longOptions, NULL);
 
             // Detect the end of the options.
             if (c == -1)
@@ -61,6 +65,12 @@ class OptionParser {
                 case 'h':
                     usage();
                     exit(0);
+                case 'v':
+                    logPolicy = "VERBOSE";
+                    break;
+                case 256:
+                    logPolicy = optarg;
+                    break;
                 case '?':
                 default:
                     // getopt_long already printed an error message.
@@ -79,6 +89,10 @@ class OptionParser {
             << std::endl
             << "scripts/failovertest.py, which kills LogCabin servers in the "
             << "meantime."
+            << std::endl
+            << std::endl
+            << "This program is subject to change (it is not part of "
+            << "LogCabin's stable API)."
             << std::endl
             << std::endl
 
@@ -108,12 +122,33 @@ class OptionParser {
             << std::endl
             << "                                 "
             << "operations [default: 10s]"
+            << std::endl
+
+            << "  -v, --verbose                  "
+            << "Same as --verbosity=VERBOSE"
+            << std::endl
+
+            << "  --verbosity=<policy>           "
+            << "Set which log messages are shown."
+            << std::endl
+            << "                                 "
+            << "Comma-separated LEVEL or PATTERN@LEVEL rules."
+            << std::endl
+            << "                                 "
+            << "Levels: SILENT ERROR WARNING NOTICE VERBOSE."
+            << std::endl
+            << "                                 "
+            << "Patterns match filename prefixes or suffixes."
+            << std::endl
+            << "                                 "
+            << "Example: Client@NOTICE,Test.cc@SILENT,VERBOSE."
             << std::endl;
     }
 
     int& argc;
     char**& argv;
     std::string cluster;
+    std::string logPolicy;
     uint64_t timeout;
 };
 
@@ -163,6 +198,9 @@ int
 main(int argc, char** argv)
 {
     OptionParser options(argc, argv);
+    LogCabin::Client::Debug::setLogPolicy(
+        LogCabin::Client::Debug::logPolicyFromString(
+            options.logPolicy));
     Cluster cluster(options.cluster);
     Tree tree = cluster.getTree();
     tree.setTimeout(options.timeout);