+++ Examples/ReconfigureTest.cc
@@ -19,6 +19,7 @@
 #include <string>
 
 #include <LogCabin/Client.h>
+#include <LogCabin/Debug.h>
 
 namespace {
 
@@ -39,16 +40,19 @@ class OptionParser {
         , argv(argv)
         , cluster("logcabin:5254")
         , fatal(false)
+        , logPolicy("")
     {
         while (true) {
             static struct option longOptions[] = {
                {"cluster",  required_argument, NULL, 'c'},
                {"fatal",  no_argument, NULL, 'f'},
-               {"seed",  required_argument, NULL, 's'},
                {"help",  no_argument, NULL, 'h'},
+               {"seed",  required_argument, NULL, 's'},
+               {"verbose",  no_argument, NULL, 'v'},
+               {"verbosity",  required_argument, NULL, 256},
                {0, 0, 0, 0}
             };
-            int c = getopt_long(argc, argv, "c:fhs:", longOptions, NULL);
+            int c = getopt_long(argc, argv, "c:fhs:v", longOptions, NULL);
 
             // Detect the end of the options.
             if (c == -1)
@@ -66,6 +70,12 @@ class OptionParser {
                     break;
                 case 's':
                     srand(static_cast<unsigned int>(atoi(optarg)));
+                    break;
+                case 'v':
+                    logPolicy = "VERBOSE";
+                    break;
+                case 256:
+                    logPolicy = optarg;
                     break;
                 case '?':
                 default:
@@ -88,6 +98,10 @@ class OptionParser {
             << "subsets of its"
             << std::endl
             << "initial configuration."
+            << std::endl
+            << std::endl
+            << "This program is subject to change (it is not part of "
+            << "LogCabin's stable API)."
             << std::endl
             << std::endl
 
@@ -118,6 +132,26 @@ class OptionParser {
 
             << "  -s <num>, --seed=<num>         "
             << "Random seed [default: 1]"
+            << std::endl
+
+            << "  -v, --verbose                  "
+            << "Same as --verbosity=VERBOSE"
+            << std::endl
+
+            << "  --verbosity=<policy>           "
+            << "Set which log messages are shown."
+            << std::endl
+            << "                                 "
+            << "Comma-separated LEVEL or PATTERN@LEVEL rules."
+            << std::endl
+            << "                                 "
+            << "Levels: SILENT ERROR WARNING NOTICE VERBOSE."
+            << std::endl
+            << "                                 "
+            << "Patterns match filename prefixes or suffixes."
+            << std::endl
+            << "                                 "
+            << "Example: Client@NOTICE,Test.cc@SILENT,VERBOSE."
             << std::endl;
     }
 
@@ -125,6 +159,7 @@ class OptionParser {
     char**& argv;
     std::string cluster;
     bool fatal;
+    std::string logPolicy;
 };
 
 void
@@ -187,6 +222,9 @@ int
 main(int argc, char** argv)
 {
     OptionParser options(argc, argv);
+    LogCabin::Client::Debug::setLogPolicy(
+        LogCabin::Client::Debug::logPolicyFromString(
+            options.logPolicy));
     Cluster cluster(options.cluster);
 
     std::pair<uint64_t, Configuration> current =